+++ OmniKit/PumpManager/OmnipodPumpManager.swift
@@ -113,7 +113,7 @@ public class OmnipodPumpManager: RileyLinkPumpManager {
             return nil
         }
 
-        let deviceProvider = RileyLinkDeviceManager(autoConnectIDs: connectionManagerState.autoConnectIDs)
+        let deviceProvider = RileyLinkBluetoothDeviceProvider(autoConnectIDs: connectionManagerState.autoConnectIDs)
 
         self.init(state: state, rileyLinkDeviceProvider: deviceProvider)
