--- RileyLinkBLEKit/RileyLinkConnectionManager.swift
@@ -1,10 +1,0 @@
-//
-//  RileyLinkConnectionManager.swift
-//  RileyLinkBLEKit
-//
-//  Created by Pete Schwamb on 8/16/18.
-//  Copyright © 2018 Pete Schwamb. All rights reserved.
-//
-
-import Foundation
-
